requirements.txt
albumentations==1.2.0
beautifulsoup4==4.12.2
dominate==2.7.0
ipdb==0.13.9
ipython==8.13.2
matplotlib==3.3.4
numpy==1.13.1
opencv_python_headless==4.2.0.34
openpyxl==3.0.9
pandas==1.1.5
Pillow==9.5.0
pycocotools==2.0.4
requests==2.25.1
retrying==1.3.4
scikit_image==0.17.2
scikit_learn==1.2.2
scipy==1.5.4
torch==1.10.1
torchvision==0.11.2
tqdm==4.19.9
ttach==0.0.3
visdom==0.1.8.9
wandb==0.12.21
